# Improves memchunk read performance.