This PR improves speed of read from memchunk by:

- inlining the moveNext function
- use a unrolled version of varint decoder (no for loop)
- pass around the context and account for stats at the end
- increase the bufio size to 64k since most block will not fit in 4k.

At the same time I also remove all dependencies of the stats package in the chunkenc package.

Final result:

```
➜ benchstat before.txt after.txt
name                                old time/op    new time/op    delta
_MemChunkIterator/entry_forward-16    17.9ms ± 8%    13.0ms ± 3%  -27.25%  (p=0.000 n=10+9)
_MemChunkIterator/sample-16           20.0ms ± 9%    15.6ms ± 2%  -21.97%  (p=0.000 n=10+9)

name                                old alloc/op   new alloc/op   delta
_MemChunkIterator/entry_forward-16    5.12kB ± 7%    3.70kB ± 8%  -27.69%  (p=0.000 n=9+8)
_MemChunkIterator/sample-16           4.07kB ±49%    2.97kB ±58%     ~     (p=0.143 n=10+10)

name                                old allocs/op  new allocs/op  delta
_MemChunkIterator/entry_forward-16      21.0 ± 0%      18.0 ± 0%  -14.29%  (p=0.000 n=10+10)
_MemChunkIterator/sample-16             21.0 ± 0%      18.0 ± 0%  -14.29%  (p=0.000 n=10+10)
```

20 to 30% boost.

Signed-off-by: Cyril Tovena <cyril.tovena@gmail.com>